// static/src/models/api.ts
/* eslint-disable */
/* tslint:disable */
// @ts-nocheck
/*
 * ---------------------------------------------------------------
 * ## THIS FILE WAS GENERATED VIA SWAGGER-TYPESCRIPT-API        ##
 * ##                                                           ##
 * ## AUTHOR: acacode                                           ##
 * ## SOURCE: https://github.com/acacode/swagger-typescript-api ##
 * ---------------------------------------------------------------
 */

export interface ApiErrorResponse {
  code?: number;
  message?: string;
}

export interface ApiItemResponseModelBrand {
  code?: number;
  data?: ModelBrand;
  message?: string;
}

export interface ApiItemResponseModelColor {
  code?: number;
  data?: ModelColor;
  message?: string;
}

export interface ApiItemResponseModelFilament {
  code?: number;
  data?: ModelFilament;
  message?: string;
}

export interface ApiItemResponseModelType {
  code?: number;
  data?: ModelType;
  message?: string;
}

export interface ApiItemResponseV1MetaData {
  code?: number;
  data?: V1MetaData;
  message?: string;
}

export interface ApiListResponseModelBrand {
  code?: number;
  data?: ModelBrand[];
  message?: string;
  pager?: ApiPager;
}

export interface ApiListResponseModelColor {
  code?: number;
  data?: ModelColor[];
  message?: string;
  pager?: ApiPager;
}

export interface ApiListResponseModelFilament {
  code?: number;
  data?: ModelFilament[];
  message?: string;
  pager?: ApiPager;
}

export interface ApiListResponseModelRecord {
  code?: number;
  data?: ModelRecord[];
  message?: string;
  pager?: ApiPager;
}

export interface ApiListResponseModelStatistic {
  code?: number;
  data?: ModelStatistic[];
  message?: string;
  pager?: ApiPager;
}

export interface ApiListResponseModelType {
  code?: number;
  data?: ModelType[];
  message?: string;
  pager?: ApiPager;
}

export interface ApiPager {
  page?: number;
  size?: number;
  total?: number;
}

export interface GormDeletedAt {
  time?: string;
  /** Valid is true if Time is not NULL */
  valid?: boolean;
}

export interface ModelBrand {
  created_at?: string;
  deleted_at?: GormDeletedAt;
  id?: number;
  name?: string;
  updated_at?: string;
}

export interface ModelColor {
  created_at?: string;
  deleted_at?: GormDeletedAt;
  id?: number;
  name?: string;
  rgba?: string;
  updated_at?: string;
}

export interface ModelFilament {
  brand?: ModelBrand;
  brand_id?: number;
  color?: ModelColor;
  color_id?: number;
  created_at?: string;
  deleted_at?: GormDeletedAt;
  id?: number;
  price?: number;
  quantity?: number;
  type?: ModelType;
  type_id?: number;
  updated_at?: string;
}

export interface ModelRecord {
  brand?: ModelBrand;
  brand_id?: number;
  color?: ModelColor;
  color_id?: number;
  created_at?: string;
  deleted_at?: GormDeletedAt;
  id?: number;
  kind?: string;
  price?: number;
  quantity?: number;
  type?: ModelType;
  type_id?: number;
  updated_at?: string;
}

export interface ModelStatistic {
  kind?: string;
  title?: string;
  values?: ModelStatisticValue[];
}

export interface ModelStatisticValue {
  children?: ModelStatisticValue[];
  name?: string;
  value?: number;
}

export interface ModelType {
  created_at?: string;
  deleted_at?: GormDeletedAt;
  id?: number;
  major?: string;
  minor?: string;
  /** ignore all, just generated by major-minor */
  name?: string;
  updated_at?: string;
}

export interface V1MetaData {
  brands?: ModelBrand[];
  colors?: ModelColor[];
  types?: ModelType[];
}